move instruction compatability test to runtime tests
