+++ runtime/vara/src/integration_tests.rs
@@ -26,6 +26,11 @@ use frame_support::{
     },
 };
 use frame_system::pallet_prelude::BlockNumberFor;
+use gear_core::gas_metering::CustomConstantCostRules;
+use gear_wasm_instrument::{
+    parity_wasm::elements::{Instruction, Module},
+    Rules,
+};
 use sp_consensus_babe::{
     digests::{PreDigest, SecondaryPlainPreDigest},
     Slot, BABE_ENGINE_ID,
@@ -860,4 +865,172 @@ fn fungible_api_works() {
                 EXISTENTIAL_DEPOSIT
             );
         });
+}
+
+#[test]
+fn process_costs_are_same() {
+    let vara_schedule = pallet_gear::Schedule::<super::Runtime>::default();
+    let wasm_schedule = gear_core::gas_metering::Schedule::default();
+
+    assert_eq!(vara_schedule.process_costs(), wasm_schedule.process_costs());
+}
+
+fn all_measured_instructions() -> Vec<Instruction> {
+    use gear_wasm_instrument::parity_wasm::elements::{BlockType, BrTableData, Instruction::*};
+
+    let default_table_data = BrTableData {
+        table: Default::default(),
+        default: 0,
+    };
+
+    // A set of instructions weights for which the Gear provides.
+    // Instruction must not be removed (!), but can be added.
+    vec![
+        End,
+        Unreachable,
+        Return,
+        Else,
+        I32Const(0),
+        I64Const(0),
+        Block(BlockType::NoResult),
+        Loop(BlockType::NoResult),
+        Nop,
+        Drop,
+        I32Load(0, 0),
+        I32Load8S(0, 0),
+        I32Load8U(0, 0),
+        I32Load16S(0, 0),
+        I32Load16U(0, 0),
+        I64Load(0, 0),
+        I64Load8S(0, 0),
+        I64Load8U(0, 0),
+        I64Load16S(0, 0),
+        I64Load16U(0, 0),
+        I64Load32S(0, 0),
+        I64Load32U(0, 0),
+        I32Store(0, 0),
+        I32Store8(0, 0),
+        I32Store16(0, 0),
+        I64Store(0, 0),
+        I64Store8(0, 0),
+        I64Store16(0, 0),
+        I64Store32(0, 0),
+        Select,
+        If(BlockType::NoResult),
+        Br(0),
+        BrIf(0),
+        Call(0),
+        GetLocal(0),
+        SetLocal(0),
+        TeeLocal(0),
+        GetGlobal(0),
+        SetGlobal(0),
+        CurrentMemory(0),
+        CallIndirect(0, 0),
+        BrTable(default_table_data.into()),
+        I32Clz,
+        I64Clz,
+        I32Ctz,
+        I64Ctz,
+        I32Popcnt,
+        I64Popcnt,
+        I32Eqz,
+        I64Eqz,
+        I64ExtendSI32,
+        I64ExtendUI32,
+        I32WrapI64,
+        I32Eq,
+        I64Eq,
+        I32Ne,
+        I64Ne,
+        I32LtS,
+        I64LtS,
+        I32LtU,
+        I64LtU,
+        I32GtS,
+        I64GtS,
+        I32GtU,
+        I64GtU,
+        I32LeS,
+        I64LeS,
+        I32LeU,
+        I64LeU,
+        I32GeS,
+        I64GeS,
+        I32GeU,
+        I64GeU,
+        I32Add,
+        I64Add,
+        I32Sub,
+        I64Sub,
+        I32Mul,
+        I64Mul,
+        I32DivS,
+        I64DivS,
+        I32DivU,
+        I64DivU,
+        I32RemS,
+        I64RemS,
+        I32RemU,
+        I64RemU,
+        I32And,
+        I64And,
+        I32Or,
+        I64Or,
+        I32Xor,
+        I64Xor,
+        I32Shl,
+        I64Shl,
+        I32ShrS,
+        I64ShrS,
+        I32ShrU,
+        I64ShrU,
+        I32Rotl,
+        I64Rotl,
+        I32Rotr,
+        I64Rotr,
+    ]
+}
+
+fn default_wasm_module() -> Module {
+    let simple_wat = r#"
+    (module
+        (import "env" "memory" (memory 1))
+        (export "handle" (func $handle))
+        (export "init" (func $init))
+        (func $handle)
+        (func $init)
+    )"#;
+    Module::from_bytes(
+        wabt::Wat2Wasm::new()
+            .validate(false)
+            .convert(simple_wat)
+            .expect("failed to parse module"),
+    )
+    .expect("module instantiation failed")
+}
+
+// This test must never fail during local development/release.
+//
+// The instruction set in the test mustn't be changed. Test checks
+// whether no instruction weight was removed from Rules, so backward
+// compatibility is reached.
+#[test]
+fn instructions_backward_compatibility() {
+    let vara_schedule = pallet_gear::Schedule::<super::Runtime>::default();
+    let wasm_schedule = gear_core::gas_metering::Schedule::default();
+
+    let schedule_rules = vara_schedule.rules(&default_wasm_module());
+    let wasm_instrument_schedule_rules = wasm_schedule.rules(&default_wasm_module());
+
+    let custom_cost_rules = CustomConstantCostRules::default();
+
+    all_measured_instructions().iter().for_each(|i| {
+        assert!(schedule_rules.instruction_cost(i).is_some());
+        assert_eq!(
+            schedule_rules.instruction_cost(i),
+            wasm_instrument_schedule_rules.instruction_cost(i)
+        );
+        assert!(custom_cost_rules.instruction_cost(i).is_some());
+    });
 }
